Change weight type from double to int (#9)

+++ mtsp-vrp-c/src/mtsp-vrp-c.cpp
@@ -5,7 +5,7 @@
 #include <chrono>
 #include <xtensor/xadapt.hpp>
 
-int solve_mtsp_vrp(size_t numberOfAgents, size_t numberOfNodes, const int* start_positions, const int* end_positions, const double* weights, int timeout,
+int solve_mtsp_vrp(size_t numberOfAgents, size_t numberOfNodes, const int* start_positions, const int* end_positions, const int* weights, int timeout,
     double* lowerBound, double* upperBound, int* paths, size_t* pathOffsets)
 {
     if (numberOfAgents == 0 || numberOfNodes < 2 || numberOfAgents * 2 > numberOfNodes)
